fix: Update import statement path (grouping refactor leftover)
