+++ src/ui/headings/PageHeading.jsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import SubHeading from "./SubHeading";
-import ResetURLButton from "./buttons/ResetURLButton";
-import RefreshButton from "./buttons/RefreshButton";
+import ResetURLButton from "../buttons/ResetURLButton";
+import RefreshButton from "../buttons/RefreshButton";
 
 const Container = styled.div`
   background-color: var(--color-grey-0);
